+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - [2551](https://github.com/FuelLabs/fuel-core/pull/2551): Enhanced the DA compressed block header to include block id.
+- [2572](https://github.com/FuelLabs/fuel-core/pull/2572): Enhanced the DA compressed block header to include tx id commitment.
 
 ## [Version 0.41.0]